Refactors sensor and monitoring config code and API
This separates out transforms and the monitoring config class into their
own modules, which necessitates a consolidation and refactoring of the API
of some of the functions in the `utils` module.
