+++ lib/tag.js
@@ -51,7 +51,9 @@ function WirelessTag(tagManager, tagData) {
         this.callAPI = platform.callAPI;
         this.log = platform.log;
     }
-    u.setObjProperties(this, roTagProps, rwTagProps);
+    u.defineOnChangeProperty(this, 'data', 'data');
+    roTagProps.forEach((p) => u.defineLinkedProperty(this, p, 'data', true));
+    rwTagProps.forEach((p) => u.defineLinkedProperty(this, p, 'data', false));
     this.data = tagData || {};
     this.on('_data', this.bounceUpdateLoop.bind(this));
 }
@@ -248,14 +250,10 @@ WirelessTag.prototype.discoverSensors = function() {
 
 function createSensor(tag, sensorType) {
     let sensor = new WirelessTagSensor(tag, sensorType);
-    let req = WirelessTagSensor.loadMonitoringConfig(sensor).then(
-        (config) => {
-            try {
-                sensor.monitoringConfig(config);
-                return sensor;
-            } finally {
-                tag.emit('discover', sensor);
-            }
+    let req = sensor.monitoringConfig().update().then(
+        () => {
+            tag.emit('discover', sensor);
+            return sensor;
         },
         (error) => {
             tag.emit('discover', sensor);
