+++ lib/tagmanager.js
@@ -45,7 +45,9 @@ function WirelessTagManager(platform, data) {
     this.wirelessTagPlatform = platform;
     this.errorHandler = platform ? platform.errorHandler : u.defaultHandler;
     this.callAPI = platform ? platform.callAPI : undefined;
-    u.setObjProperties(this, roMgrProps, rwMgrProps);
+    u.defineOnChangeProperty(this, 'data', 'data');
+    roMgrProps.forEach((p) => u.defineLinkedProperty(this, p, 'data', true));
+    rwMgrProps.forEach((p) => u.defineLinkedProperty(this, p, 'data', false));
     this.data = data;
 }
 util.inherits(WirelessTagManager, EventEmitter);
