+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "wirelesstags",
-  "version": "0.6.2",
+  "version": "0.7.0-beta.0",
   "description": "Interface to the Wireless Sensor Tags platform (http://wirelesstag.net)",
   "main": "index.js",
   "scripts": {
